Extract model price sync module
