+++ internal/service/admin.go
@@ -4,14 +4,11 @@ import (
 	"crypto/rand"
 	"encoding/base32"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"log"
-	"math"
 	"net"
 	"net/http"
-	"os"
 	"reflect"
 	"sort"
 	"strconv"
@@ -22,33 +19,12 @@ import (
 	"github.com/awsl-project/maxx/internal/adapter/provider"
 	"github.com/awsl-project/maxx/internal/domain"
 	"github.com/awsl-project/maxx/internal/event"
+	"github.com/awsl-project/maxx/internal/modelpricesync"
 	"github.com/awsl-project/maxx/internal/pricing"
 	"github.com/awsl-project/maxx/internal/repository"
 	"github.com/awsl-project/maxx/internal/systemsettingcache"
 	"github.com/awsl-project/maxx/internal/version"
 )
-
-const defaultModelPriceSyncSourceURL = "https://raw.githubusercontent.com/BerriAI/litellm/main/model_prices_and_context_window.json"
-
-const defaultModelPriceSyncSourceCode = "litellm"
-
-type modelPriceSyncSource struct {
-	Code       string
-	Name       string
-	DefaultURL string
-	EnvURL     string
-	Convert    func(modelID string, item liteLLMModelPrice) (*domain.ModelPrice, bool)
-}
-
-var modelPriceSyncSources = map[string]modelPriceSyncSource{
-	defaultModelPriceSyncSourceCode: {
-		Code:       defaultModelPriceSyncSourceCode,
-		Name:       "LiteLLM",
-		DefaultURL: defaultModelPriceSyncSourceURL,
-		EnvURL:     "MAXX_MODEL_PRICE_SYNC_SOURCE_URL",
-		Convert:    convertLiteLLMModelPrice,
-	},
-}
 
 // ProviderAdapterRefresher is an interface for refreshing provider adapters
 // Implemented by Router to receive notifications when providers change
@@ -2020,22 +1996,6 @@ type ModelPriceChange struct {
 	After  *domain.ModelPrice `json:"after"`
 }
 
-type liteLLMModelPrice struct {
-	Mode                                      string   `json:"mode"`
-	LitellmProvider                           string   `json:"litellm_provider"`
-	InputCostPerToken                         *float64 `json:"input_cost_per_token"`
-	OutputCostPerToken                        *float64 `json:"output_cost_per_token"`
-	CacheReadInputTokenCost                   *float64 `json:"cache_read_input_token_cost"`
-	CacheCreationInputTokenCost               *float64 `json:"cache_creation_input_token_cost"`
-	CacheCreationInputTokenCostAbove1hr       *float64 `json:"cache_creation_input_token_cost_above_1hr"`
-	InputCostPerImageToken                    *float64 `json:"input_cost_per_image_token"`
-	OutputCostPerImageToken                   *float64 `json:"output_cost_per_image_token"`
-	InputCostPerTokenAbove200kTokens          *float64 `json:"input_cost_per_token_above_200k_tokens"`
-	OutputCostPerTokenAbove200kTokens         *float64 `json:"output_cost_per_token_above_200k_tokens"`
-	CacheReadInputTokenCostAbove200kTokens    *float64 `json:"cache_read_input_token_cost_above_200k_tokens"`
-	CacheCreationInputTokenCostAbove200kToken *float64 `json:"cache_creation_input_token_cost_above_200k_tokens"`
-}
-
 // RecalculateCosts recalculates cost for all attempts using the current price table
 // and updates the parent requests' cost accordingly (with streaming batch processing)
 func (s *AdminService) RecalculateCosts() (*RecalculateCostsResult, error) {
@@ -2261,7 +2221,7 @@ func (s *AdminService) ResetModelPricesToDefaults() ([]*domain.ModelPrice, error
 // and returns the pending DB changes without applying them.
 func (s *AdminService) PreviewModelPricesFromExternalSource(sourceCode ...string) (*SyncModelPricesResult, error) {
 	source := firstModelPriceSyncSource(sourceCode)
-	prices, resolvedSource, sourceURL, err := fetchExternalModelPrices(source)
+	prices, resolvedSource, sourceURL, err := modelpricesync.Fetch(source)
 	if err != nil {
 		return nil, err
 	}
@@ -2272,7 +2232,7 @@ func (s *AdminService) PreviewModelPricesFromExternalSource(sourceCode ...string
 // applies missing or changed prices. Custom DB-only prices are preserved.
 func (s *AdminService) SyncModelPricesFromExternalSource(sourceCode ...string) (*SyncModelPricesResult, error) {
 	source := firstModelPriceSyncSource(sourceCode)
-	prices, resolvedSource, sourceURL, err := fetchExternalModelPrices(source)
+	prices, resolvedSource, sourceURL, err := modelpricesync.Fetch(source)
 	if err != nil {
 		return nil, err
 	}
@@ -2336,116 +2296,6 @@ func (s *AdminService) syncModelPrices(sourcePrices []*domain.ModelPrice, source
 		result.Prices = prices
 	}
 	return result, nil
-}
-
-func fetchExternalModelPrices(sourceCode string) ([]*domain.ModelPrice, modelPriceSyncSource, string, error) {
-	source, err := resolveModelPriceSyncSource(sourceCode)
-	if err != nil {
-		return nil, source, "", err
-	}
-	sourceURL := os.Getenv(source.EnvURL)
-	if sourceURL == "" {
-		sourceURL = source.DefaultURL
-	}
-
-	client := &http.Client{Timeout: 30 * time.Second}
-	resp, err := client.Get(sourceURL)
-	if err != nil {
-		return nil, source, sourceURL, fmt.Errorf("fetch model price source %q: %w", source.Code, err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return nil, source, sourceURL, fmt.Errorf("fetch model price source %q: unexpected status %d", source.Code, resp.StatusCode)
-	}
-
-	var raw map[string]liteLLMModelPrice
-	if err := json.NewDecoder(resp.Body).Decode(&raw); err != nil {
-		return nil, source, sourceURL, fmt.Errorf("decode model price source %q: %w", source.Code, err)
-	}
-
-	prices := make([]*domain.ModelPrice, 0, len(raw))
-	for modelID, item := range raw {
-		price, ok := source.Convert(modelID, item)
-		if ok {
-			prices = append(prices, price)
-		}
-	}
-	sort.Slice(prices, func(i, j int) bool { return prices[i].ModelID < prices[j].ModelID })
-	return prices, source, sourceURL, nil
-}
-
-func resolveModelPriceSyncSource(sourceCode string) (modelPriceSyncSource, error) {
-	code := strings.TrimSpace(strings.ToLower(sourceCode))
-	if code == "" {
-		code = defaultModelPriceSyncSourceCode
-	}
-	source, ok := modelPriceSyncSources[code]
-	if !ok {
-		return modelPriceSyncSource{}, fmt.Errorf("unsupported model price sync source %q", sourceCode)
-	}
-	return source, nil
-}
-
-func convertLiteLLMModelPrice(modelID string, item liteLLMModelPrice) (*domain.ModelPrice, bool) {
-	if modelID == "" || strings.HasPrefix(modelID, "sample_") {
-		return nil, false
-	}
-
-	input := costPerTokenToMicroPerM(item.InputCostPerToken)
-	output := costPerTokenToMicroPerM(item.OutputCostPerToken)
-	imageInput := costPerTokenToMicroPerM(item.InputCostPerImageToken)
-	imageOutput := costPerTokenToMicroPerM(item.OutputCostPerImageToken)
-	if input == 0 && output == 0 && imageInput == 0 && imageOutput == 0 {
-		return nil, false
-	}
-
-	price := &domain.ModelPrice{
-		ModelID:                modelID,
-		InputPriceMicro:        input,
-		OutputPriceMicro:       output,
-		CacheReadPriceMicro:    costPerTokenToMicroPerM(item.CacheReadInputTokenCost),
-		Cache5mWritePriceMicro: costPerTokenToMicroPerM(item.CacheCreationInputTokenCost),
-		Cache1hWritePriceMicro: costPerTokenToMicroPerM(item.CacheCreationInputTokenCostAbove1hr),
-		ImageInputPriceMicro:   imageInput,
-		ImageOutputPriceMicro:  imageOutput,
-	}
-	if price.Cache1hWritePriceMicro == 0 {
-		price.Cache1hWritePriceMicro = price.Cache5mWritePriceMicro
-	}
-
-	inputAbove := costPerTokenToMicroPerM(item.InputCostPerTokenAbove200kTokens)
-	outputAbove := costPerTokenToMicroPerM(item.OutputCostPerTokenAbove200kTokens)
-	if inputAbove > input && outputAbove > output && input > 0 && output > 0 {
-		price.Has1MContext = true
-		price.Context1MThreshold = 200000
-		price.InputPremiumNum, price.InputPremiumDenom = ratio(inputAbove, input)
-		price.OutputPremiumNum, price.OutputPremiumDenom = ratio(outputAbove, output)
-	}
-
-	return price, true
-}
-
-func costPerTokenToMicroPerM(v *float64) uint64 {
-	if v == nil || *v <= 0 {
-		return 0
-	}
-	return uint64(math.Round(*v * 1_000_000_000_000))
-}
-
-func ratio(numerator, denominator uint64) (uint64, uint64) {
-	if denominator == 0 || numerator == 0 {
-		return 0, 0
-	}
-	g := gcd(numerator, denominator)
-	return numerator / g, denominator / g
-}
-
-func gcd(a, b uint64) uint64 {
-	for b != 0 {
-		a, b = b, a%b
-	}
-	return a
 }
 
 func cloneModelPrice(price *domain.ModelPrice) *domain.ModelPrice {
